requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
openai==1.58.1
openpyxl==3.1.5
jinja2==3.1.5
aiosmtplib==3.0.2
python-dotenv==1.0.1
